feat: improve text component
